% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{log_summed_exps}
\usage{
log_summed_exps(to_sum)
}
\arguments{
\item{to_sum}{A numeric vector.}
}
\value{
Natural logarithm of the sum of exponentiated values of vector
}
\description{
log_summed_exps
}
\examples{
log_summed_exps(c(1, 2, 3))
}
